+++ src/idunno.Bluesky/Unspecced/Server/GetPostThreadV2.cs
@@ -1,0 +1,115 @@
+// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
+using System.Text;
+
+using idunno.AtProto;
+using idunno.AtProto.Authentication;
+using idunno.Bluesky.Unspecced;
+using idunno.Bluesky.Unspecced.Model;
+
+using Microsoft.Extensions.Logging;
+
+namespace idunno.Bluesky;
+
+public static partial class BlueskyServer
+{
+    /// <summary>
+    /// Get posts in a thread. It is based in an anchor post at any depth of the tree, and returns posts above it (recursively resolving the parent, without further branching to their replies)
+    /// and below it
+    /// </summary>
+    /// <param name="anchor">Reference <see cref="AtUri"/> to post record. This is the anchor post, and the thread will be built around it. It can be any post in the tree, not necessarily a root post.</param>
+    /// <param name="above">Flag indicating whether to include parents above the anchor.</param>
+    /// <param name="below">How many levels to include below the anchor.</param>
+    /// <param name="branchingFactor">Maximum of replies to include at each level of the thread, except for the direct replies to the anchor, which are (NOTE: currently, during unspecced phase) all returned (NOTE: later they might be paginated).</param>
+    /// <param name="sort">The sort order for the thread.</param>
+    /// <param name="service">The <see cref="Uri"/> of the service to retrieve the profile from.</param>
+    /// <param name="accessCredentials">The <see cref="AccessCredentials"/> used to authenticate to <paramref name="service"/>.</param>
+    /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
+    /// <param name="loggerFactory">An instance of <see cref="ILoggerFactory"/> to use to create a logger.</param>
+    /// <param name="subscribedLabelers">An optional list of <see cref="Did"/>s of labelers to retrieve labels applied to the account.</param>
+    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>The task object representing the asynchronous operation.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="anchor"/>, <paramref name="accessCredentials"/>, <paramref name="service"/> or <paramref name="httpClient" /> are <see langword="null"/>.</exception>
+    [UnconditionalSuppressMessage("Trimming",
+        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
+        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
+    [UnconditionalSuppressMessage("AOT",
+        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
+        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
+    [Experimental("BSKYUnspecced", UrlFormat = "https://bluesky.idunno.dev/docs/unspecced.html")]
+    [SuppressMessage("Globalization", "CA1308:Normalize strings to uppercase", Justification = "javascript require lowercase")]
+    public static async Task<AtProtoHttpResult<PostThreadV2>> GetPostThreadV2(
+        AtUri anchor,
+        bool? above,
+        int? below,
+        int? branchingFactor,
+        string? sort,
+        Uri service,
+        AccessCredentials accessCredentials,
+        HttpClient httpClient,
+        Action<AtProtoCredential>? onCredentialsUpdated = null,
+        ILoggerFactory? loggerFactory = default,
+        IEnumerable<Did>? subscribedLabelers = null,
+        CancellationToken cancellationToken = default)
+    {
+        ArgumentNullException.ThrowIfNull(anchor);
+
+        ArgumentNullException.ThrowIfNull(accessCredentials);
+        ArgumentNullException.ThrowIfNull(service);
+        ArgumentNullException.ThrowIfNull(httpClient);
+
+        StringBuilder queryStringBuilder = new($"anchor={Uri.EscapeDataString(anchor.ToString())}");
+        if (above.HasValue)
+        {
+            queryStringBuilder.Append(CultureInfo.InvariantCulture, $"&above={above.Value.ToString().ToLowerInvariant()}");
+        }
+        if (below.HasValue)
+        {
+            queryStringBuilder.Append(CultureInfo.InvariantCulture, $"&below={below.Value}");
+        }
+        if (branchingFactor.HasValue)
+        {
+            queryStringBuilder.Append(CultureInfo.InvariantCulture, $"&branchingFactor={branchingFactor.Value}");
+        }
+        if (sort is not null)
+        {
+            queryStringBuilder.Append(CultureInfo.InvariantCulture, $"&sort={Uri.EscapeDataString(sort)}");
+        }
+
+        string queryString = queryStringBuilder.ToString();
+
+        BlueskyHttpClient<GetPostThreadV2Response> request = new(AppViewProxy, loggerFactory);
+
+        AtProtoHttpResult<GetPostThreadV2Response> result = await request.Get(
+            service,
+            $"xrpc/app.bsky.unspecced.getPostThreadV2?{queryString}",
+            credentials: accessCredentials,
+            httpClient: httpClient,
+            jsonSerializerOptions: BlueskyJsonSerializerOptions,
+            onCredentialsUpdated: onCredentialsUpdated,
+            cancellationToken: cancellationToken).ConfigureAwait(false);
+
+        if (result.Succeeded)
+        {
+            return new AtProtoHttpResult<PostThreadV2>(
+                new PostThreadV2(result.Result),
+                statusCode: result.StatusCode,
+                httpResponseHeaders: result.HttpResponseHeaders,
+                atErrorDetail: result.AtErrorDetail,
+                rateLimit: result.RateLimit);
+        }
+        else
+        {
+            return new AtProtoHttpResult<PostThreadV2>(
+                null,
+                statusCode: result.StatusCode,
+                httpResponseHeaders: result.HttpResponseHeaders,
+                atErrorDetail: result.AtErrorDetail,
+                rateLimit: result.RateLimit);
+        }
+    }
+}