+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Added support for muting actors reposts and quoteposts, with the addition of scopes to the `MuteActor` method to specify the type of mute. See [Add repost and quotepost-only mutes](https://github.com/bluesky-social/atproto/pull/5118).
 * Added optional `FailureCode` property to JobStatus class to provide machine-readable failure codes for video processing jobs. Known values are defined in the `FailureCodes` class. See [https://github.com/bluesky-social/atproto/pull/5283](Add video job failure codes).
 * Added new, undocumented, `Uploading` and `Encoding` states to the `JobState` enum to reflect the discoved video processing states.
+* Added implementation of the unspecced `GetPostThreadV2` and `GetPostThreadV2Other` apis.
 
 ## Breaking Changes